Implement value-type for raw_trigger_data ROOT conversion
The `raw_trigger_data` type uses shared_ptr and vectors of shared_ptr
for each of the three record types. Status of streaming shared_ptrs in
ROOT trees is not 100% clear, and there is no hard requirement for
downstream use of shared_ptr ownership or semantics in RTD (it only
appears useful at the RHD to RTD merge stage).

To simplify serialization for offline use, create a new RRawTriggerData
class which holds the same data as raw_trigger_data but without use
of shared_ptr. Only provide const member functions and returns with
direct construction so that it can obey "read only" semantics. Provide
a ROOT dictionary binding to enable it to be stored in TTree.

Update rtd2asroot example to use the new class, reading in .data.gz
RTD files and converting the stream of raw_trigger_data to a
RRawTriggerData branch on an output TTree.