Add root startup script for testing
Root needs the path to the SNemoRawDataObj headers and have loaded
the library in order to Read/Write data, and interact with the
Raw Data Objects in the interpreter.

Add a macro/script that can be passed to `root` to set this up
automatically for users testing from a local build.